% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_plot.R
\name{act_plot}
\alias{act_plot}
\title{trelliscope auxillary function}
\usage{
act_plot(id, id_Nday, act_ori, act_ind, act_all, act_max, band, ori = F,
  lw = T)
}
\arguments{
\item{id}{individual ID.}

\item{id_Nday}{the ith day observation of the individual.}

\item{act_ori}{day observation.}

\item{act_ind}{individual mean observation.}

\item{act_all}{global mean observation.}

\item{act_max}{the upper limit of activity (y-axis).}

\item{band}{smoothing band paramter.}
}
\value{
individual to day observations.
}
\description{
Plot day observations.
}
\keyword{internal}
