% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.harmonic.R
\name{fft.harmonic}
\alias{fft.harmonic}
\title{FFT auxillary function}
\usage{
fft.harmonic(Xk)
}
\arguments{
\item{Xk}{the trelliscope data frame containing two variables: subject ID and activity}
}
\value{
Organized FFT results from top significant frequencies.
}
\description{
Take FFT input and reorganize it showing top significant frequencies.
}
\keyword{internal}
