% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_plot.R
\name{ind_plot}
\alias{ind_plot}
\title{trelliscope auxillary function}
\usage{
ind_plot(id, act_ind, act_all, band, ori = TRUE, lw = FALSE)
}
\arguments{
\item{id}{individual ID.}

\item{act_ind}{individual mean observation.}

\item{act_all}{global mean observation.}

\item{band}{smoothing band paramter.}
}
\value{
individual to day observations.
}
\description{
plot individual observations.
}
\keyword{internal}
