% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_to_day.R
\name{ind_to_day}
\alias{ind_to_day}
\title{trelliscope auxillary function}
\usage{
ind_to_day(x, df)
}
\arguments{
\item{x}{covariates data for individuals to be merged.}

\item{df}{day observation dataset.}
}
\value{
individual to day observations.
}
\description{
Take individual characteristics and map it to individual-day observations. For example, individual A/B is 40/45 
years old and has 4/2 day observations in dataset D. Therefore, mapping age to dataset D generates 
40, 40, 40, 40, 45, 45 for day 1,2,3,4/1,2 observation of A/B respectively.
}
\keyword{internal}
