\name{lis3}
\docType{data}
\alias{lis3}
\title{An example of individual activity data}
\description{
  'lis3' is a data list consisting of three matrices, each giving
  activity data for one individual. Each column of the matrix is 
  one-day observation, and here the physical activity (PA) is 
  measured every one minute, so the matrix is 1440 by 'nday'. 
  'lis3' is also a named list, the name of which is the individual ID.
}
\usage{lis3}
\format{A data list of activity data for 3 individuals; each inidivdual data set is a 1440 by nday matrix.}

\keyword{datasets}