\name{pa3}
\docType{data}
\alias{pa3}
\title{An example of reformated individual activity data}
\description{
  'pa3' is in tbl_df format. It has 13 observations for 3 individuals, and the variables are "ID", "ID_Nday" (the ith day observation for an individual), and activity. The activity variable is an embedded list with each element consisting of a vector of one-day observation.
}
\usage{data(pa3)}
\format{tbl_df}

\keyword{datasets}