\name{var3}
\docType{data}
\alias{var3}
\title{Demographic information for individuals in datasets lis3 and pa3}
\description{
  'var3' is a data frame consisting of ID, gender, and age information for the 3 individuals in datasets 'lis3' and 'pa3'. 'var3' is used as an input to generate trelliscope panels for visualization so that one can filter based on age and gender.
}
\usage{data(var3)}
\format{a data frame consisting of three variables: ID, gender and age}

\keyword{datasets}